Fix getInfo in InfoThread
